new header move newsletter to footer (#573)
* new header move newsletter to footer

* Remove yc logo for now

* Fix resource section

* Fix resource section

---------
